refactor: svg 로고 크기 속성 제거 (#18)
