+++ src/components/layout/Header.tsx
@@ -1,4 +1,4 @@
-import Logo from '@/assets/logo/headerlogo.svg';
+import Logo from '@/assets/icons/logo.svg';
 import ProfileImage from '@/assets/user/userprofile.svg';
 import Link from 'next/link';
 import {ROUTES} from '@/constants/routes';
@@ -7,7 +7,7 @@ export const Header = () => {
   return (
     <header className='bg-background sticky top-0 z-50 flex items-center justify-between px-6 py-4 shadow-sm'>
       <Link href={ROUTES.HOMEPAGE}>
-        <Logo />
+        <Logo className='h-5 w-55' />
       </Link>
 
       <div className='h-9 w-9 overflow-hidden rounded-full'>
